trying task 2

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,9 @@
 #Kill the process menow
 exec { 'killmenow':
-  command => 'pkill -f killmenow'
+  command     => 'pkill -f killmenow',
+  path        => ['/bin', '/usr/bin'],
+  onlyif      => 'pgrep -f killmenow',
+  logoutput   => true,
+  refreshonly => true,
 }
